prov/shm: Touch all buffer pages after DSA page fault
DSA page faults will now always trigger touching all of the pages in the
descriptor buffer (as it is done in dsa_touch_buffer_pages()), instead of
just the individual page causing the fault (as was done in
dsa_handle_page_fault()).

Previously, the default behavior was to touch the individual faulting page,
and if the user enabled the FI_SHM_ENABLE_DSA_PAGE_TOUCH flag, then all the
rest of the pages in the descriptor buffer were also touched. This commit
removes this flag, and now always touches all the pages instead.

Signed-off-by: Jorge Cabrera <jorge.cabrera@intel.com>